Custom autocmd PublishPre instead of BufReadPost
I've developed another Vim plug-in that I'm planning to release soon,
which automatically creates and keeps up to date a global tags file and
uses this tags file to dynamically syntax highlight all user-defined
symbols. The syntax highlighting isn't performed on BufReadPost because
that slows Vim down way too much, but I nevertheless wanted the
highlighting to apply in the source code I publish. Now I've added the
following line to the other plug-in and all is well :-)

    autocmd! User PublishPre :HighlightTags

PS. You can see the highlighted function names on the page
http://peterodding.com/code/vim/profile/plugin/publish.vim